% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMic.R
\name{openMic}
\alias{openMic}
\alias{html_dependency_open_mic}
\title{openMic}
\usage{
openMic(location = c("bottom", "top"))

html_dependency_open_mic(location)
}
\arguments{
\item{location}{location where closed captions should appear: "top" or "bottom"}
}
\value{
An \code{htmltools::tagList()} with the tile view dependencies, or an
\code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}}.
}
\description{
openMic gives you a way to have live captioning appear over your slides!
}
\section{Functions}{
\itemize{
\item \code{openMic}: Adds closed captioning to your xaringan slides.

\item \code{html_dependency_open_mic}: Returns an \code{\link[htmltools:htmlDependency]{htmltools::htmlDependency()}} with the openMic
dependencies. Most users will want to use \code{openMic()}.
}}

\section{Usage}{
 To add live captioning to your xaringan presentation, add the
following code chunk to your slides' R Markdown file.\if{html}{\out{<div class="markdown">}}\preformatted{```\{r open-mic, echo=FALSE\}
openMic::openMic()
```
}\if{html}{\out{</div>}}
}

